refactoring using Backend class
